+++ packages/keyring-api/src/api/v2/index.ts
@@ -1,0 +1,7 @@
+export type * from './keyring';
+export * from './keyring-capabilities';
+export * from './keyring-type';
+export * from './create-account';
+export * from './export-account';
+export * from './private-key';
+export * from './wrapper';