+++ packages/keyring-snap-bridge/src/SnapKeyring.ts
@@ -1,10 +1,5 @@
-/* eslint-disable @typescript-eslint/no-redundant-type-constituents */
-// This rule seems to be triggering a false positive. Possibly eslint is not
-// inferring the `EthMethod`, `BtcMethod`, and `InternalAccount` types correctly.
-
 import type { TypedTransaction } from '@ethereumjs/tx';
 import { TransactionFactory } from '@ethereumjs/tx';
-import type { ExtractEventPayload } from '@metamask/base-controller';
 import type { TypedDataV1, TypedMessage } from '@metamask/eth-sig-util';
 import { SignTypedDataVersion } from '@metamask/eth-sig-util';
 import type {
@@ -39,8 +34,14 @@ import {
   type InternalAccount,
 } from '@metamask/keyring-internal-api';
 import { KeyringInternalSnapClient } from '@metamask/keyring-internal-snap-client';
+import {
+  type GetSelectedAccountsResponse,
+  GetSelectedAccountsRequestStruct,
+  SnapManageAccountsMethod,
+} from '@metamask/keyring-snap-sdk';
 import type { AccountId, JsonRpcRequest } from '@metamask/keyring-utils';
 import { strictMask } from '@metamask/keyring-utils';
+import type { ExtractEventPayload } from '@metamask/messenger';
 import type { SnapId } from '@metamask/snaps-sdk';
 import { type Snap } from '@metamask/snaps-utils';
 import { assert, mask, object, string } from '@metamask/superstruct';
@@ -51,7 +52,6 @@ import {
   KnownCaipNamespace,
   toCaipChainId,
 } from '@metamask/utils';
-import { EventEmitter } from 'events';
 import { v4 as uuid } from 'uuid';
 
 import { transformAccount } from './account';
@@ -74,6 +74,7 @@ import type {
   SnapKeyringEvents,
   SnapKeyringMessenger,
 } from './SnapKeyringMessenger';
+import { SNAP_KEYRING_NAME } from './SnapKeyringMessenger';
 import type { SnapMessage } from './types';
 import { SnapMessageStruct } from './types';
 import {
@@ -154,10 +155,15 @@ function normalizeAccountAddress(account: KeyringAccount): string {
 /**
  * Keyring bridge implementation to support Snaps.
  */
-export class SnapKeyring extends EventEmitter {
+export class SnapKeyring {
   static type: string = SNAP_KEYRING_TYPE;
 
   type: string;
+
+  // Name and state are required for modular initialisation.
+  name: typeof SNAP_KEYRING_NAME = SNAP_KEYRING_NAME;
+
+  state = null;
 
   /**
    * Messenger to dispatch requests to the Snaps controller.
@@ -177,6 +183,11 @@ export class SnapKeyring extends EventEmitter {
     account: KeyringAccount;
     snapId: SnapId;
   }>;
+
+  /**
+   * Mapping between Snap IDs and the selected accounts.
+   */
+  readonly #selectedAccounts: Map<SnapId, AccountId[]>;
 
   /**
    * Mapping between request IDs and their deferred promises.
@@ -225,7 +236,6 @@ export class SnapKeyring extends EventEmitter {
     callbacks: SnapKeyringCallbacks;
     isAnyAccountTypeAllowed?: boolean;
   }) {
-    super();
     this.type = SnapKeyring.type;
     this.#messenger = messenger;
     this.#snapClient = new KeyringInternalSnapClient({ messenger });
@@ -234,6 +244,7 @@ export class SnapKeyring extends EventEmitter {
     this.#options = new SnapIdMap();
     this.#callbacks = callbacks;
     this.#isAnyAccountTypeAllowed = isAnyAccountTypeAllowed;
+    this.#selectedAccounts = new Map();
   }
 
   /**
@@ -314,6 +325,7 @@ export class SnapKeyring extends EventEmitter {
 
     // Potentially migrate the account.
     const account = transformAccount(newAccountFromEvent);
+    const address = normalizeAccountAddress(account);
 
     // The `AnyAccountType.Account` generic account type is allowed only during
     // development, so we check whether it's allowed before continuing.
@@ -324,10 +336,21 @@ export class SnapKeyring extends EventEmitter {
       throw new Error(`Cannot create generic account '${account.id}'`);
     }
 
+    // This is idempotent, so we need to check whether the account already exists
+    // and that the right Snap is trying to "create" it again.
+    const accountEntry = this.#accounts.get(snapId, account.id);
+    if (
+      accountEntry &&
+      normalizeAccountAddress(accountEntry.account) === address
+    ) {
+      // NOTE: We are not checking account object equality here. If a Snap
+      // re-send this event with different account data, we will ignore it.
+      return null;
+    }
+
     // The UI still uses the account address to identify accounts, so we need
     // to block the creation of duplicate accounts for now to prevent accounts
     // from being overwritten.
-    const address = normalizeAccountAddress(account);
     if (await this.#callbacks.addressExists(address)) {
       throw new Error(`Account address '${address}' already exists`);
     }
@@ -491,6 +514,21 @@ export class SnapKeyring extends EventEmitter {
       },
     );
     return null;
+  }
+
+  /**
+   * Handle a Get Selected Accounts method call from a Snap.
+   *
+   * @param snapId - Snap ID.
+   * @param message - Method call message.
+   * @returns The selected accounts.
+   */
+  async #handleGetSelectedAccounts(
+    snapId: SnapId,
+    message: SnapMessage,
+  ): Promise<GetSelectedAccountsResponse> {
+    assert(message, GetSelectedAccountsRequestStruct);
+    return this.#selectedAccounts.get(snapId) ?? [];
   }
 
   /**
@@ -695,6 +733,10 @@ export class SnapKeyring extends EventEmitter {
 
       case `${KeyringEvent.AccountTransactionsUpdated}`: {
         return this.#handleAccountTransactionsUpdated(snapId, message);
+      }
+
+      case `${SnapManageAccountsMethod.GetSelectedAccounts}`: {
+        return this.#handleGetSelectedAccounts(snapId, message);
       }
 
       default:
@@ -1494,6 +1536,48 @@ export class SnapKeyring extends EventEmitter {
   }
 
   /**
+   * Update the in-memory selected accounts map.
+   *
+   * @param accounts - The accounts to update the map with.
+   */
+  #updateSelectedAccountsMap(accounts: AccountId[]): void {
+    const selectedAccounts = this.#selectedAccounts;
+    selectedAccounts.clear();
+    for (const account of accounts) {
+      const snapId = this.#accounts.getSnapId(account);
+      if (!snapId) {
+        continue;
+      }
+      const snapAccounts = selectedAccounts.get(snapId) ?? [];
+      snapAccounts.push(account);
+      selectedAccounts.set(snapId, snapAccounts);
+    }
+  }
+
+  /**
+   * Set the selected accounts.
+   *
+   * @param accounts - The accounts to set as selected.
+   */
+  async setSelectedAccounts(accounts: AccountId[]): Promise<void> {
+    this.#updateSelectedAccountsMap(accounts);
+    const entries = [...this.#selectedAccounts.entries()];
+    await Promise.all(
+      entries.map(async ([snapId, accountIds]) => {
+        try {
+          await this.#snapClient
+            .withSnapId(snapId)
+            .setSelectedAccounts(accountIds);
+        } catch (error: any) {
+          console.error(
+            `Failed to set selected accounts for ${snapId} snap: '${error.message}'`,
+          );
+        }
+      }),
+    );
+  }
+
+  /**
    * Get the Snap associated with the given Snap ID.
    *
    * @param snapId - Snap ID.
@@ -1531,6 +1615,35 @@ export class SnapKeyring extends EventEmitter {
     );
   }
 
+  #transformToInternalAccount(
+    account: KeyringAccount,
+    snapId: SnapId,
+  ): InternalAccount {
+    const snap = this.#getSnapMetadata(snapId);
+
+    return {
+      ...account,
+      // TODO: Do not convert the address to lowercase.
+      //
+      // This is a workaround to support the current UI which expects the
+      // account address to be lowercase. This workaround should be removed
+      // once we migrated the UI to use the account ID instead of the account
+      // address.
+      //
+      // NOTE: We convert the address only for EVM accounts, see
+      // `normalizeAccountAddress`.
+      address: normalizeAccountAddress(account),
+      metadata: {
+        name: '',
+        importTime: 0,
+        keyring: {
+          type: this.type,
+        },
+        ...(snap !== undefined && { snap }),
+      },
+    };
+  }
+
   /**
    * Return an internal account object for a given address.
    *
@@ -1538,41 +1651,26 @@ export class SnapKeyring extends EventEmitter {
    * @returns An internal account object for the given address.
    */
   getAccountByAddress(address: string): InternalAccount | undefined {
-    const accounts = this.listAccounts();
-    return accounts.find(({ address: accountAddress }) =>
+    const accounts = [...this.#accounts.values()];
+    const account = accounts.find(({ account: { address: accountAddress } }) =>
       equalsIgnoreCase(accountAddress, address),
     );
+
+    return account
+      ? this.#transformToInternalAccount(account.account, account.snapId)
+      : undefined;
   }
 
   /**
    * List all Snap keyring accounts.
+   * This method is expensive on mobile devices and could takes tens or hundreds of milliseconds to complete.
+   * Use with caution.
    *
    * @returns An array containing all Snap keyring accounts.
    */
   listAccounts(): InternalAccount[] {
-    return [...this.#accounts.values()].map(({ account, snapId }) => {
-      const snap = this.#getSnapMetadata(snapId);
-      return {
-        ...account,
-        // TODO: Do not convert the address to lowercase.
-        //
-        // This is a workaround to support the current UI which expects the
-        // account address to be lowercase. This workaround should be removed
-        // once we migrated the UI to use the account ID instead of the account
-        // address.
-        //
-        // NOTE: We convert the address only for EVM accounts, see
-        // `normalizeAccountAddress`.
-        address: normalizeAccountAddress(account),
-        metadata: {
-          name: '',
-          importTime: 0,
-          keyring: {
-            type: this.type,
-          },
-          ...(snap !== undefined && { snap }),
-        },
-      };
-    });
+    return [...this.#accounts.values()].map(({ account, snapId }) =>
+      this.#transformToInternalAccount(account, snapId),
+    );
   }
 }